Add build rule for examples to generate executable binaries. Fix javadoc errors in examples.
