+++ examples/src/main/java/com/google/instrumentation/examples/trace/BasicTracing.java
@@ -11,21 +11,20 @@
  * limitations under the License.
  */
 
-package com.google.instrumentation.trace;
+package com.google.instrumentation.examples.trace;
 
-/** Example showing how to directly create a child {@link Span} and add annotations. */
-public final class MultiSpansTracing {
+import com.google.instrumentation.trace.Span;
+import com.google.instrumentation.trace.Tracer;
+
+/** Example showing how to create a {@link Span} and add annotations. */
+public final class BasicTracing {
   // Per class Tracer.
   private static final Tracer tracer = Tracer.getTracer();
 
   private static void doWork() {
-    Span rootSpan = tracer.spanBuilder(null, "MyRootSpan").startSpan();
-    rootSpan.addAnnotation("Annotation to the root Span before child is created.");
-    Span childSpan = tracer.spanBuilder(rootSpan, "MyChildSpan").startSpan();
-    childSpan.addAnnotation("Annotation to the child Span");
-    childSpan.end();
-    rootSpan.addAnnotation("Annotation to the root Span after child is ended.");
-    rootSpan.end();
+    Span span = tracer.spanBuilder(null, "MyRootSpan").startSpan();
+    span.addAnnotation("This annotation is added directly to the span.");
+    span.end();
   }
 
   /** Main method. */
