+++ examples/src/main/java/com/google/instrumentation/examples/trace/MultiSpansContextTracing.java
@@ -11,9 +11,11 @@
  * limitations under the License.
  */
 
-package com.google.instrumentation.trace;
+package com.google.instrumentation.examples.trace;
 
 import com.google.instrumentation.common.NonThrowingCloseable;
+import com.google.instrumentation.trace.Span;
+import com.google.instrumentation.trace.Tracer;
 
 /**
  * Example showing how to create a child {@link Span}, install it to the current context and add
